# Conflicts:
#	app/src/main/java/com/worldventures/dreamtrips/core/module/ApiModule.java
#	app/src/main/java/com/worldventures/dreamtrips/core/navigation/Route.java
#	app/src/main/java/com/worldventures/dreamtrips/modules/trips/view/fragment/TripDetailsFragment.java
#	app/src/main/res/values/styles.xml

+++ app/src/main/java/com/worldventures/dreamtrips/modules/common/presenter/LaunchActivityPresenter.java
@@ -193,6 +193,7 @@ public class LaunchActivityPresenter extends ActivityPresenter<LaunchActivityPre
     private void clearTempDirectory() {
         snappyRepository.removeAllUploadTasks();
         File directory = new File(com.kbeanie.imagechooser.api.FileUtils.getDirectory(PickImageDelegate.FOLDERNAME));
+        if (!directory.exists()) return;
         try {
             FileUtils.cleanDirectory(context, directory);
         } catch (IOException e) {
